test: nestjs Account 예시에 testcontainers 기반 e2e 테스트 추가
실제로 기동 가능한 앱으로 완성하기 위해 누락되어 있던 공유 모듈
(common/generate-id, common/generate-error-response,
database/transaction-manager, outbox/*)과 app-module/main.ts,
package.json/tsconfig/jest 설정을 추가했다.

test/account.e2e-spec.ts는 @testcontainers/postgresql로 실제 Postgres를
띄우고 8개 엔드포인트의 요청 형식·응답·에러 케이스(404/400/검증 실패 등)를
모두 supertest로 검증한다. 테스트 작성 과정에서 실제 버그 3개를 발견해
수정했다: ownerId 컬럼이 char(32)로 고정폭 패딩되던 문제, amount 컬럼이
bigint라 JS에서 문자열로 반환되던 문제, GetTransactionsQuery의 page/take에
실제 기본값이 없어 생략 시 검증 실패하던 문제.

+++ implementations/nestjs/examples/src/account/infrastructure/entity/transaction.entity.ts
@@ -11,7 +11,7 @@ export class TransactionEntity {
   @Column()
   type: string
 
-  @Column('bigint')
+  @Column('int')
   amount: number
 
   @Column({ type: 'char', length: 3 })
